Factor out bucket segment routing code to SegmentRoutingUtil
